fix: move new settings to components directory
